Consolidate platformSetup: remove redundant index.website.ts
index.native.ts and index.ts were both no-ops after removing
Performance. Move the web implementation from index.website.ts
into index.ts and delete the .website variant.
